Add support for MongoDB 5.1+ to mc_worker_api
The MongoDB protocol has evolved and the protocol that this driver used
before this commit has been removed in MongoDB 5.1. This commit makes it
possible to use the module mc_worker_api (that handles a connection to a
single MongoDB server) with MongoDB 5.1+ by configuring the driver with
`application:set_env(mongodb, use_legacy_protocol, false)`. This
functionality can be tested by running
`./test/test_without_legacy_protocol.sh`.

This is the first step in making the whole driver compatible with
MongoDB 5.1+.

The commit also includes some fixes to Dialyzer warnings that existed
even before the commit (there are still a few warnings left).